Add basic error handling

+++ src/posts/posts.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { PostEntity } from './post.entity';
@@ -20,15 +20,27 @@ export class PostsService {
   };
 
   async getById(id: number) {
-    return await this.postRepo.findOneBy({ id });
+    const post = await this.postRepo.findOneBy({ id });
+    if(!post) {
+      throw new HttpException(`Post with ID: ${id} not found`, HttpStatus.NOT_FOUND)
+    };
+    return post;
   };
 
   async update(id: number, dto: UpdatePostDto) {
+    const post = await this.getById(id);
+    if(!post) {
+      throw new HttpException(`Post with ID: ${id} not found`, HttpStatus.NOT_FOUND)
+    };
     await this.postRepo.update(id, dto);
     return await this.getById(id);
   };
 
   async delete(id: number) {
+    const post = await this.getById(id);
+    if(!post) {
+      throw new HttpException(`Post with ID: ${id} not found`, HttpStatus.NOT_FOUND)
+    };
     await this.postRepo.delete({ id });
     return { message: `Post with ID: ${id} successfully deleted` };
   };
